# backend/api.yaml
openapi: 3.0.3
info:
  title: Placement Portal Application API
  description: |
    REST API documentation for Placement Portal Application.
    The system provides role-based services for Admin, Company, and Student users.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
...
components:
  securitySchemes:

    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - JWT: []
paths:
...
  /api/admin/companies:

    get:

      tags:

        - Admin

      summary:

        Get all companies


      parameters:

        - name: search

          in: query

          schema:

            type: string


      responses:

        "200":

          description:

            Company list
  /api/admin/companies/{user_id}/approve:

    put:

      tags:

        - Admin

      summary:

        Approve company


      parameters:

        - name: user_id

          in: path

          required: true

          schema:

            type: integer


      responses:

        "200":

          description:

            Company approved
  /api/admin/companies/{user_id}/reject:

    delete:

      tags:

        - Admin

      summary:

        Reject company


      parameters:

        - name: user_id

          in: path

          required: true

          schema:

            type: integer


      responses:

        "200":

          description:

            Company rejected
...